Add new "deletion request" APIs. (#984)
* Refactor UserPolicy for readability.

* Add policy & full response to user deletion endpoint.

* Expose 'deletion_requested_at' on users.

* Add "deletion request" resource.

* Fix issue where AuthorizationException would be returned as 500s.

This just casts this exception to the HttpException equivalent so it doesn't return as a 500.

* Adds test coverage for the new deletion endpoints.

* Fix up mismatched docblocks.

+++ app/Http/Controllers/UserController.php
@@ -178,15 +178,19 @@ class UserController extends Controller
      * Delete a user resource.
      * DELETE /users/:id
      *
-     * @param $id - User ID
+     * @param User $user
      * @return \Illuminate\Http\Response
      * @throws NotFoundHttpException
      */
-    public function destroy($id)
+    public function destroy(User $user)
     {
-        $user = User::findOrFail($id);
+        $this->authorize('delete', $user);
+
+        // @see: UserObserver@deleting
         $user->delete();
 
-        return $this->respond('No Content.');
+        // We'll return the deleted user since that may be helpful
+        // for some applications to update their local state.
+        return $this->item($user);
     }
 }
